Refactor all views into seperate files
