+++ task_manager/urls.py
@@ -16,10 +16,8 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from tasks import views
 from django.conf.urls import handler404
-from django.urls import re_path as url
-from tasks.views import home
+import tasks.views as views
 
 tasks = 'teams'
 task_manager = 'teams'
@@ -33,19 +31,15 @@ urlpatterns = [
     path('password/', views.PasswordView.as_view(), name='password'),
     path('profile/', views.ProfileUpdateView.as_view(), name='profile'),
     path('sign_up/', views.SignUpView.as_view(), name='sign_up'),
-    # path('create_team/', views.create_team, name='create_team'),
     path('create_team/', views.CreateTeamView.as_view(), name='create_team'),
     path('team/<int:team_id>', views.show_team, name='show_team'),
     path('team/<int:team_id>/invite/', views.invite, name='invite'),
     path('invitations/', views.list_invitations, name='list_invitations'),
     path('invitations/accept/<int:invitation_id>/', views.accept_invitation, name='accept_invitation'),
     path('invitations/decline/<int:invitation_id>/', views.decline_invitation, name='decline_invitation'),
-    # path('create_task/<int:team_id>', views.create_task, name="create_task" ),
     path('create_task/<int:team_id>', views.CreateTaskView.as_view(), name="create_task" ),
     path('team/edit_team/<int:team_id>', views.TeamUpdateView.as_view(), name='edit_team'),
-    # path('edit_task/<int:task_id>', views.edit_task, name='edit_task'),
     path('edit_task/<int:task_id>', views.EditTaskView.as_view(), name='edit_task'),
-    # path('delete_task/<int:task_id>', views.delete_task, name = 'delete_task'),
     path('delete_task/<int:task_id>', views.DeleteTaskView.as_view(), name = 'delete_task'),
     path('assign_member_to_task/<int:task_id>/<int:user_id>', views.assign_member_to_task, name= "assign_member_to_task"),
     path('notifications', views.notifications, name="notifications"),
@@ -53,7 +47,6 @@ urlpatterns = [
     path('toggle_archive/<int:task_id>', views.toggle_task_archive, name='toggle_archive'),
     path('leaderboard/<int:team_id>', views.leaderboard_view, name='leaderboard'),
     path('seen_notification/<int:notification_id>', views.seen_notification, name="seen_notification"),
-    # path('delete/<int:team_id>', views.team_delete, name='team_delete'),
     path('delete/<int:team_id>', views.DeleteTeamView.as_view(), name='team_delete'),
 ]
