+++ tasks/views/__init__.py
@@ -1,0 +1,8 @@
+from .account_views import *
+from .authentication_views import *
+from .team_views import *
+from .task_views import *
+from .notification_views import *
+from .static_views import *
+from .invitation_views import *
+from .dashboard_views import *
