+++ tasks/views/authentication_views.py
@@ -1,0 +1,50 @@
+"""Authenticated related views."""
+from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth import login, logout
+from django.shortcuts import redirect, render
+from django.views import View
+from django.urls import reverse
+from tasks.forms import LogInForm
+from .mixins import LoginProhibitedMixin
+
+class LogInView(LoginProhibitedMixin, View):
+    """Display login screen and handle user login."""
+
+    http_method_names = ['get', 'post']
+    redirect_when_logged_in_url = settings.REDIRECT_URL_WHEN_LOGGED_IN
+
+    def get(self, request):
+        """Display log in template."""
+
+        self.next = request.GET.get('next') or ''
+        return self.render()
+
+    def post(self, request):
+        """Handle log in attempt."""
+
+        form = LogInForm(request.POST)
+        self.next = request.POST.get('next') or settings.REDIRECT_URL_WHEN_LOGGED_IN
+        user = form.get_user()
+        if user is not None:
+            login(request, user)
+            messages.add_message(request, messages.SUCCESS, "You have successfully Logged in!")
+            return redirect(self.next)
+        messages.add_message(request, messages.ERROR, "The credentials provided were invalid!")
+        return self.render()
+
+    def render(self):
+        """Render log in template with blank log in form."""
+
+        form = LogInForm()
+        return render(self.request, 'log_in.html', {'form': form, 'next': self.next})
+    
+def log_out(request):
+    """Log out the current user"""
+
+    logout(request)
+    messages.add_message(request, messages.SUCCESS, "You have signed out")
+    return redirect('home')
+
+
+
